requirements.txt
streamlit==1.49.0
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.16
langchain-groq==1.1.2
chromadb==1.0.20
beautifulsoup4==4.12.3
pandas==2.3.2
python-dotenv==1.1.1
pytest==9.0.2
